annotation_tool/Annotation-Interface/requirements.txt
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
ansi2html==1.8.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli==1.0.9
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
comm==0.2.0
ConfigArgParse==1.5.3
contourpy==1.1.0
cryptography==41.0.5
cycler==0.10.0
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.9
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dpcpp-cpp-rt==2023.2.0
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.17.1
ffmpeg==1.4
filelock==3.13.1
fire==0.5.0
Flask==2.2.5
Flask-Cors==4.0.0
fonttools==4.40.0
fqdn==1.5.1
gmpy2==2.1.5
gunicorn==21.1.0
idna==2.10
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.2.0
intel-cmplr-lic-rt==2023.2.0
intel-opencl-rt==2023.2.0
intel-openmp==2023.2.0
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-bbox-widget==0.5.0
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter-events==0.6.3
jupyter_packaging==0.12.3
jupyter_server==2.10.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lit==17.0.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mkl==2023.2.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
mpmath==1.2.1
mutagen==1.46.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.4.0
nbformat==5.5.0
nest-asyncio==1.5.8
networkx==3.2.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open3d==0.15.2
opencv-python==4.7.0.72
overrides==7.4.0
packaging==23.2
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.3
platformdirs==3.11.0
plotly==5.15.0
proglog==0.1.10
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.6
pycparser==2.21
pycryptodomex==3.18.0
pygame==2.5.0
pyglet==1.5.27
Pygments==2.16.1
PyOpenGL==3.1.1a1
pyOpenSSL==23.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytube==15.0.0
pytz==2023.3
PyVirtualDisplay==3.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.0
scikit-learn==1.2.2
scipy==1.10.1
segment-anything==1.0
Send2Trash==1.8.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
supervision==0.11.1
svg.path==6.3
sympy==1.12
tbb==2021.10.0
tenacity==8.2.2
tensorboardX==2.6.2
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.13.0
trimesh==3.22.2
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==2.2.3
wget==3.2
wheel==0.41.2
widgetsnbextension==4.0.9
y-py==0.6.2
ypy-websocket==0.8.4
yt-dlp==2023.11.16
zipp==3.17.0
